Remove libp2p demos
